+++ src/main/java/com/skycat/mystical/accessor/MinecraftServerTimerAccess.java
@@ -1,0 +1,6 @@
+package com.skycat.mystical.accessor;
+
+
+public interface MinecraftServerTimerAccess {
+    void mystical_setTimer(long ticksUntilNight);
+}