+++ src/main/java/com/skycat/mystical/mixin/MinecraftServerTimer.java
@@ -1,7 +1,7 @@
 package com.skycat.mystical.mixin;
 
 import com.skycat.mystical.Mystical;
-import com.skycat.mystical.accessors.MinecraftServerTimerAccess;
+import com.skycat.mystical.accessor.MinecraftServerTimerAccess;
 import net.minecraft.server.MinecraftServer;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
@@ -11,7 +11,6 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 import java.util.function.BooleanSupplier;
 
-// CREDIT: Daomephsta#0044 for help on fabric discord (and fabric bot too)
 @Mixin(MinecraftServer.class)
 public abstract class MinecraftServerTimer implements MinecraftServerTimerAccess {
     @Unique