+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 - Internal changes
 - Added a spell HUD for players with the mod installed on the client
+- Added /mystical spell delete all
+- Added some translations
+- illusionersReplaceEvokers checks for havens now
+- Added /mystical credits
 - Added help commands
-- illusionersReplaceEvokers accounts for havens
-- /mystical spell delete all
-- /mystical credits